requirements.txt
Pillow
opencv-python
extract_msg
urllib3
requests
dnspython
email
dkimpy
cryptography